clownxc/Project/vegetables_classify/requirements.txt
tensorflow==2.7.0
matplotlib~=3.5.2
numpy~=1.21.6
requests~=2.28.1
flask~=2.1.3
keras~=2.7.0
pillow~=9.2.0